feat(example) follow ctaut change 1.database add UpdatedHeight、UpdateHistoryHeights 2. update dependency

+++ examples/ctaut/burn.go
@@ -54,11 +54,11 @@ func burnCTAUT(identifier abelian.AutId) {
 			HideValue:     true,
 		}
 
-		if metadata.PrivacyType == abelian.AutPrivacyTypeUnlimited {
+		if metadata.PrivacyType == uint8(abelian.AutPrivacyTypeUnlimited) {
 			// unlimited privacy type, all recipients are public / hidden
-		} else if metadata.PrivacyType == abelian.AutPrivacyTypeLimitedPublic {
+		} else if metadata.PrivacyType == uint8(abelian.AutPrivacyTypeLimitedPublic) {
 			recipient.HideValue = false
-		} else if metadata.PrivacyType == abelian.AutPrivacyTypeLimitedHidden {
+		} else if metadata.PrivacyType == uint8(abelian.AutPrivacyTypeLimitedHidden) {
 			recipient.HideValue = true
 		} else {
 			panic("unsupported privacy type")
@@ -115,11 +115,11 @@ func burnCTAUT(identifier abelian.AutId) {
 			Value:         selectedValue - targetTokenValue,
 			HideValue:     false,
 		}
-		if metadata.PrivacyType == abelian.AutPrivacyTypeUnlimited {
+		if metadata.PrivacyType == uint8(abelian.AutPrivacyTypeUnlimited) {
 			// unlimited privacy type, all recipients are public / hidden
-		} else if metadata.PrivacyType == abelian.AutPrivacyTypeLimitedPublic {
+		} else if metadata.PrivacyType == uint8(abelian.AutPrivacyTypeLimitedPublic) {
 			recipient.HideValue = false
-		} else if metadata.PrivacyType == abelian.AutPrivacyTypeLimitedHidden {
+		} else if metadata.PrivacyType == uint8(abelian.AutPrivacyTypeLimitedHidden) {
 			recipient.HideValue = true
 		} else {
 			panic("unsupported privacy type")
